2D shapes for Ships  were created, also GameStage and GameController were deleted and WelcomeController and WelcomeStage were renamed
