+++ src/main/java/com/example/navalbattlejjc/view/GameStage.java
@@ -1,4 +1,33 @@
 package com.example.navalbattlejjc.view;
 
-public class GameStage {
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+import javafx.scene.Parent;
+import javafx.fxml.FXMLLoader;
+
+import java.io.IOException;
+
+public class WelcomeStage extends Stage {
+    public WelcomeStage() throws IOException {
+        setTitle("Naval Battle");
+        setMaximized(true);
+
+        // Load the FXML file
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("game-view.fxml"));
+        Parent root = loader.load();
+
+        // Create the welcome scene and set to the stage
+        Scene welcomeScene = new Scene(root);
+        setScene(welcomeScene);
+
+        show();
+    }
+
+    public static WelcomeStage getInstance() throws IOException {
+        return WelcomeStageHolder.INSTANCE = new WelcomeStage();
+    }
+
+    private static class WelcomeStageHolder {
+        private static WelcomeStage INSTANCE;
+    }
 }
